notif [nfc]: Split out module notifTokens
Part of the reorganization described in the preceding commit.
